Promote SignbankSearchService and supporting code to DictionarySearchService
