+++ app/views/dictionary_signs/_card.html.erb
