+++ app/views/search/index.html.erb
@@ -15,21 +15,21 @@
 </div>
 <hr class="list__divider list__divider--dark">
 
-<% if @freelex_signs.any? %>
+<% if @dictionary_signs.any? %>
 <div class="grid-x cell" id="official-signs">
   <div class="cell margin-top-3 margin-bottom-0 grid-x align-middle">
     <div class="cell grid-x align-middle">
       <h2 class="cell auto">NZSL Dictionary signs</h2>
       <div class="cell shrink grid-x align-middle">
-        <%= link_to FREELEX_CONFIG[:public_host] do %>
+        <%= link_to Rails.application.config.dictionary_host do %>
           See more dictionary signs<%= inline_svg_pack_tag("media/images/external-link.svg", class: "icon icon--inline icon--royal") %>
         <% end %>
       </div>
     </div>
   </div>
   <div class="sign-grid cell">
-    <% @freelex_signs.each do |sign| %>
-      <%= render "freelex_signs/card", sign: sign %>
+    <% @dictionary_signs.each do |sign| %>
+      <%= render "dictionary_signs/card", sign: sign %>
     <% end %>
   </div>
 </div>
