+++ spec/system/search_signs_spec.rb
@@ -17,37 +17,8 @@ RSpec.describe "Searching for signs" do
   end
 
   context "with only Signbank results" do
-    around do |example|
-      original_dictionary_sign_model = Rails.application.config.dictionary_sign_model
-      Rails.application.config.dictionary_sign_model = DictionarySign
-      example.run
-    ensure
-      Rails.application.config.dictionary_sign_model = original_dictionary_sign_model
-    end
-
     before do
-      FactoryBot.create_list(:freelex_sign, 6, word: "a")
-      submit_search("a")
-    end
-
-    it "shows the expected results" do
-      # We only show a preview of official signs
-      expect(page).to have_selector "#official-signs .sign-card", count: 4
-      expect(page).to have_no_selector "#community-signs"
-    end
-  end
-
-  context "with only Freelex results" do
-    around do |example|
-      original_dictionary_sign_model = Rails.application.config.dictionary_sign_model
-      Rails.application.config.dictionary_sign_model = FreelexSign
-      example.run
-    ensure
-      Rails.application.config.dictionary_sign_model = original_dictionary_sign_model
-    end
-
-    before do
-      FactoryBot.create_list(:freelex_sign, 6, word: "a")
+      FactoryBot.create_list(:dictionary_sign, 6, word: "a")
       submit_search("a")
     end
 
@@ -61,7 +32,6 @@ RSpec.describe "Searching for signs" do
   context "with only Share results" do
     before do
       DictionarySign.unscoped.destroy_all
-      FreelexSign.unscoped.destroy_all
       FactoryBot.create_list(:sign, 5, :published, word: "a")
       submit_search("a")
     end
